schat_backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.10
django-cors-headers==3.10.1
django-mongoengine==0.4.6
djangorestframework==3.12.4
django-phonenumber-field==6.0.0
django-jsonfield==1.4.1
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
mongoengine==0.23.1
pi==0.1.2
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
whitenoise==5.3.0
cryptography==2.8